+++ packages/terra-card/src/Card.jsx
@@ -7,19 +7,35 @@ import CardBody from './CardBody';
 
 const cx = classNames.bind(styles);
 
+const CardVariants = {
+  DEFAULT: 'default',
+  RAISED: 'raised',
+};
+
+
 const propTypes = {
   /**
    * Child Nodes
    */
   children: PropTypes.node,
+  /**
+   * Sets the card variant to change the style for different use cases. One of `default`,  `raised`.
+   */
+  variant: PropTypes.oneOf([...Object.values(CardVariants)]),
+};
+
+const defaultProps = {
+  variant: CardVariants.DEFAULT,
 };
 
 const Card = ({
   children,
+  variant,
   ...customProps
 }) => {
   const cardClassNames = cx([
     'card',
+    variant,
     customProps.className,
   ]);
 
@@ -27,5 +43,6 @@ const Card = ({
 };
 
 Card.propTypes = propTypes;
+Card.defaultProps = defaultProps;
 Card.Body = CardBody;
 export default Card;