+++ packages/terra-collapsible-menu-view/tests/jest/CollapsibleMenuViewItem.test.jsx
@@ -1,22 +1,23 @@
 import React from 'react';
 import IconTrash from 'terra-icon/lib/icon/IconTrash';
 import CollapsibleMenuViewItem from '../../src/CollapsibleMenuViewItem';
+import intlContexts from './intl-context-setup';
 
 describe('CollapsibleMenuViewItem', () => {
   // Snapshot Tests
   it('should render a default component', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" />);
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
   // Props
   it('should merge custom props', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" className="Testing" />);
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" className="Testing" />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should not set isSelected on button outside of an item group', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" isSelected />);
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" isSelected />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -33,27 +34,27 @@ describe('CollapsibleMenuViewItem', () => {
   });
 
   it('should set icon prop on button', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" icon={<IconTrash />} />);
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" icon={<IconTrash />} />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should set isReversed prop on button', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" icon={<IconTrash />} isReversed />);
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" icon={<IconTrash />} isReversed />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('should render a button with icon and no text when isIconOnly is set', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" icon={<IconTrash />} isIconOnly />);
+  it('should render a button with icon and text when isIconOnly and text is set', () => {
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" icon={<IconTrash />} isIconOnly />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a menu when subMenuItems are given', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" subMenuItems={[<CollapsibleMenuViewItem text="Menu Item" key="1" />]} />);
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" subMenuItems={[<CollapsibleMenuViewItem text="Menu Item" key="1" />]} />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a disabled button when isDisabled is set', () => {
-    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" isDisabled />);
+    const wrapper = shallow(<CollapsibleMenuViewItem text="Testing" isDisabled />, intlContexts.shallowContext);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -107,7 +108,7 @@ describe('CollapsibleMenuViewItem', () => {
           text="Testing"
           subMenuItems={[<CollapsibleMenuViewItem text="Menu Item" key="1" />]}
         />
-      ), { context });
+    ), { context });
       expect(wrapper).toMatchSnapshot();
     });