+++ packages/terra-content-container/examples/test-examples/FillContentContainer.example.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import ContentContainer from '../../lib/ContentContainer';
+
+const header = <header>header</header>;
+const content = <div style={{ height: '100%', backgroundColor: '#98fb98' }}>content</div>;
+
+const container = () => (
+  <div style={{ height: '200px' }}>
+    <ContentContainer id="content-container" header={header} fill>
+      {content}
+    </ContentContainer>
+  </div>);
+
+export default container;