+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Add more z-indexes.
+
+1.1.0 - (May 2, 2018)
+------------------
+### Changed
+* Updated outdated documentation
+* Minor dependency version bump
 
 1.0.0 - (April 20, 2018)
 ------------------