+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -4,6 +4,14 @@ ChangeLog
 Unreleased
 ----------
 
+2.8.0 - (May 2, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+Added localized aria-label to ellipses "view more" button
+
 2.7.0 - (April 20, 2018)
 ------------------
 ### Changed