+++ packages/terra-doc-template/CHANGELOG.md
@@ -2,5 +2,8 @@ ChangeLog
 =========
 
 Unreleased
------------------
+----------
+
+1.0.0 - (May 2, 2018)
+------------------
 * Initial stable release